Split into headers and CPP files
  * Added compiler hints for templates
